refactor(belinvestbank): extract base64 helper
