+++ src/plugins/belinvestbank/fingerprint.js
@@ -1,3 +1,5 @@
+import { encodeUtf8ToBase64 } from './base64'
+
 // Device profiles, fingerprint generation, and mobileSdkData emulation
 // for Belinvestbank mobile API anti-fraud requirements
 
@@ -100,58 +102,6 @@ RcKU18IVYcmzCkZymo7An3zD68Pq38TGn1QcYieV8vdE18uLGUkRnFN1bqodNFu5
 9FjOp+7y/TY6Iv819Q==
 -----END CERTIFICATE-----`
 
-const BASE64_CHARS = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/'
-
-function utf8ToBytes (str) {
-  const octets = []
-  let i = 0
-  while (i < str.length) {
-    const codePoint = str.codePointAt(i)
-    let shift = 0
-    let prefix = 0
-    if (codePoint <= 0x7F) {
-      shift = 0
-      prefix = 0x00
-    } else if (codePoint <= 0x7FF) {
-      shift = 6
-      prefix = 0xC0
-    } else if (codePoint <= 0xFFFF) {
-      shift = 12
-      prefix = 0xE0
-    } else {
-      shift = 18
-      prefix = 0xF0
-    }
-    octets.push(prefix | (codePoint >> shift))
-    shift -= 6
-    while (shift >= 0) {
-      octets.push(0x80 | ((codePoint >> shift) & 0x3F))
-      shift -= 6
-    }
-    i += codePoint >= 0x10000 ? 2 : 1
-  }
-  return octets
-}
-
-function bytesToBase64 (bytes) {
-  let result = ''
-  for (let i = 0; i < bytes.length; i += 3) {
-    const a = bytes[i]
-    const b = i + 1 < bytes.length ? bytes[i + 1] : 0
-    const c = i + 2 < bytes.length ? bytes[i + 2] : 0
-    const chunk = (a << 16) | (b << 8) | c
-    result += BASE64_CHARS[(chunk >> 18) & 0x3F]
-    result += BASE64_CHARS[(chunk >> 12) & 0x3F]
-    result += i + 1 < bytes.length ? BASE64_CHARS[(chunk >> 6) & 0x3F] : '='
-    result += i + 2 < bytes.length ? BASE64_CHARS[chunk & 0x3F] : '='
-  }
-  return result
-}
-
-function utf8ToBase64 (str) {
-  return bytesToBase64(utf8ToBytes(str))
-}
-
 // Seeded PRNG (mulberry32) — deterministic random from login
 function createSeededRandom (seed) {
   let h = 0
@@ -205,7 +155,7 @@ function seededFcmToken (rand) {
   for (let i = 0; i < 22; i++) instanceId += chars[(rand() * chars.length) | 0]
   let token = 'APA91b'
   for (let i = 0; i < 134; i++) token += tokenChars[(rand() * tokenChars.length) | 0]
-  return utf8ToBase64(`${instanceId}:${token}`)
+  return encodeUtf8ToBase64(`${instanceId}:${token}`)
 }
 
 function generateRandomGeo () {
@@ -341,7 +291,7 @@ function generateMobileSdkData (fp) {
     PhoneNumber: '',
     DeviceName: fp.deviceModel
   }
-  return utf8ToBase64(JSON.stringify(sdkData, null, 2))
+  return encodeUtf8ToBase64(JSON.stringify(sdkData, null, 2))
 }
 
 export function mobileHeaders (fp, cookie) {
